requirements.txt
streamlit
torch
pandas
Pillow
ultralytics